Move resetting resp. merging markedPackages onFileStateChanged to its redux service
