+++ visualization/app/codeCharta/state/store.service.spec.ts
@@ -2,23 +2,18 @@ import "./state.module"
 import { StoreService } from "./store.service"
 import { getService, instantiateModule } from "../../../mocks/ng.mockhelper"
 import { IRootScopeService } from "angular"
-import { BlacklistItem, BlacklistType, FileSelectionState, FileState } from "../codeCharta.model"
+import { BlacklistItem, BlacklistType } from "../codeCharta.model"
 import { BlacklistAction, BlacklistActions } from "./store/fileSettings/blacklist/blacklist.actions"
-import { DEFAULT_STATE, STATE, TEST_DELTA_MAP_A, TEST_DELTA_MAP_B } from "../util/dataMocks"
+import { DEFAULT_STATE, STATE } from "../util/dataMocks"
 import { setState } from "./store/state.actions"
 import { setDynamicSettings } from "./store/dynamicSettings/dynamicSettings.actions"
 import { setMargin } from "./store/dynamicSettings/margin/margin.actions"
-import { FileStateService } from "./fileState.service"
-import { FileStateHelper } from "../util/fileStateHelper"
-import { SettingsMerger } from "../util/settingsMerger"
 import { LoadingStatusService } from "./loadingStatus.service"
 
 describe("StoreService", () => {
 	let storeService: StoreService
 	let $rootScope: IRootScopeService
 	let loadingStatusService: LoadingStatusService
-
-	let fileStates: FileState[]
 
 	beforeEach(() => {
 		restartSystem()
@@ -30,11 +25,6 @@ describe("StoreService", () => {
 
 		$rootScope = getService<IRootScopeService>("$rootScope")
 		loadingStatusService = getService<LoadingStatusService>("loadingStatusService")
-
-		fileStates = [
-			{ file: JSON.parse(JSON.stringify(TEST_DELTA_MAP_A)), selectedAs: FileSelectionState.Comparison },
-			{ file: JSON.parse(JSON.stringify(TEST_DELTA_MAP_B)), selectedAs: FileSelectionState.Reference }
-		]
 	}
 
 	function rebuildService() {
@@ -46,52 +36,6 @@ describe("StoreService", () => {
 			rebuildService()
 
 			expect(storeService["store"]).not.toBeNull()
-		})
-
-		it("should subscribe to FileStateService", () => {
-			FileStateService.subscribe = jest.fn()
-
-			rebuildService()
-
-			expect(FileStateService.subscribe).toHaveBeenCalledWith($rootScope, storeService)
-		})
-	})
-
-	describe("onFileStatesChanged", () => {
-		beforeEach(() => {
-			FileStateHelper.isPartialState = jest.fn().mockReturnValue(false)
-			FileStateHelper.getVisibleFileStates = jest.fn().mockReturnValue(fileStates)
-			SettingsMerger.getMergedFileSettings = jest.fn().mockReturnValue(DEFAULT_STATE)
-		})
-
-		it("should update store with default dynamicSettings and newFileSettings", () => {
-			storeService.onFileStatesChanged(fileStates)
-
-			expect(storeService.getState().dynamicSettings.focusedNodePath).toEqual("")
-			expect(storeService.getState().dynamicSettings.searchedNodePaths).toEqual([])
-			expect(storeService.getState().dynamicSettings.searchPattern).toEqual("")
-			expect(storeService.getState().dynamicSettings.margin).toBeNull()
-			expect(storeService.getState().dynamicSettings.colorRange).toEqual({ from: null, to: null })
-			expect(storeService.getState().fileSettings).toEqual(DEFAULT_STATE.fileSettings)
-		})
-
-		it("should call isPartialState with fileStates", () => {
-			storeService.onFileStatesChanged(fileStates)
-
-			expect(FileStateHelper.isPartialState).toHaveBeenCalledWith(fileStates)
-		})
-
-		it("should call getVisibleFileStates with fileStates", () => {
-			storeService.onFileStatesChanged(fileStates)
-
-			expect(FileStateHelper.getVisibleFileStates).toHaveBeenCalledWith(fileStates)
-		})
-
-		it("should call getMergedFileStates with visibleFiles and withUpdatedPath", () => {
-			storeService.onFileStatesChanged(fileStates)
-			const visibleFiles = [fileStates[0].file, fileStates[1].file]
-
-			expect(SettingsMerger.getMergedFileSettings).toHaveBeenCalledWith(visibleFiles, false)
 		})
 	})
